Corrected ECS notebook, added figs, simplified
Added a correction to state that my "Equilibrium Climate Sensivity" is
actually calculated only ~100-150 years after quadrupling of CO2, so
this is in fact very far from equilibrium.

Added figures of the historical and ssp585 warming.

Simplified the global warming notebooks by removing the regridding and
simplified the area-weighting to cos(lat)-weighting (equivalent but
simplier).